Improvements for effects
Values which are modified by multiplying them, are fixed. Some effects are still too fast or move strangely - I assume the next step would be checking the VectorAdd/VectorSubtract calls which have o->Position as target.
Prepared code to improve random checks on each frame.

+++ Source Main 5.2/source/ZzzObject.h
@@ -9,8 +9,8 @@ extern OBJECT       Fishs[];
 extern float EarthQuake;
 extern bool EnableShadow;
 extern float AmbientShadowAngle;
-extern int RainTarget;
-extern int RainCurrent;
+extern float RainTarget;
+extern float RainCurrent;
 
 void CopyShadowAngle(OBJECT* o, BMD* b);
 void CreateShadowAngle();
